plugin events

+++ bowery/agent/build.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 echo "Building Bowery Agent!"
-rm dev-agent
+if [ -e dev-agent ]; then
+  rm dev-agent
+fi
+
 GOROOT=/usr/lib/go GOPATH=/root/gopath go get -d
-GOROOT=/usr/lib/go GOPATH=/root/gopath go build
+GOROOT=/usr/lib/go GOPATH=/root/gopath go build -o dev-agent
